Validation post (#61)
The user can now post a comment and an image while validating a challenge. If the challenge was already validated, the previous comment and image are replaced. One can verify the validation by refreshing the Challenge screen to see its previously posted comment.
If the challenge has already been completed, the comment and image are displayed if they exist. #34 #62